move generic private key methods into their own module
